requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==1.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==6.2.0
psycopg2==2.7.3
requests==2.20.0
SQLAlchemy==1.1.13
urllib3==1.24.2
Werkzeug==0.15.3
WTForms==2.1
numpy==1.19.1
rembg==1.0.14

